django-app/requirements.txt
asgiref==3.5.2
Django
psycopg2-binary==2.9.5
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.2.0
python-dotenv==0.20.0
djangorestframework==3.14.0
Pillow
django_filter
django_advance_thumbnail
celery
redis
